Userspace Stage B3 step 12: Darwin float-argument storage, missing double parameter, libm prototypes
- GLDriver's float formatter FUN_000cdc3c called ecvt with its integer
  arguments in SysV registers (r3/r4/r5, Darwin uses r5/r6/r7), so the
  rebuilt image died with SIGBUS. It also read its own value from an
  undeclared f1 that all 17 stock callers load. Fixed with
  DarwinStorage.java (apply) and AddDoubleParams.java.
- libm imports had no signature, so every caller read the result from r3
  or an unassigned extraout_f1. SetLibmSignatures.java adds the prototypes
  (GLDriver 11, glprog 17, VA 2); ghidra2c declares them with their real
  return type.
- in_f1..f13 classified with inreg_liveness.py fpr; tables in
  inreg_liveness_fpr.tsv.
- ghidra2c: reinterpret a pointer cast of a float-cursor element, and the
  float use of that word.
- New tests: ftoa_test.c (20 values, stock == rebuilt) and glsl_libm.vert
  (the probe takes a shader file). Tests README updated.

+++ Userspace/ATIRadeonX1000GLDriver/ppc/part_004.c
@@ -1891,7 +1891,7 @@ LAB_000256f4:
       fVar2 = *(float *)(iVar23 + 0x24f4);
       fVar3 = *(float *)(iVar23 + 0x24f0);
       fVar4 = *(float *)(iVar23 + 0x24f8) + FLOAT_001aa0e8;
-      fparam_1 = (double)_sqrt((double)(fVar4 * fVar4 + fVar3 * fVar3 + fVar2 * fVar2));
+      fparam_1 = _sqrt((double)(fVar4 * fVar4 + fVar3 * fVar3 + fVar2 * fVar2));
       local_84[0] = fVar6;
       fVar6 = (float)(DOUBLE_001aa200 / fparam_1);
       local_88 = fVar4 * fVar6;
